Lots of changes in this commit: The titanium plugin now follows a structure closer to that of what we saw in Gears (some code was copied over to get the plugin entry points / bug fixes etc).
For now the plugin is loading on it's own in Safari (you can see the debug outs) until we try to initialize the EditorObject using NPN_CreateObject, where it crashes. Probably something to do with the use of g_browser_funcs, but not sure yet.

On the bright side, at least we can test our plugin now instead of bootstrapping all of gears for our POC ;)