Clean up Process and remove deadlocks.
All pipe output is now buffered and buffers are protected
by mutexes. If an onread is attached after process exit it
is always fired if there was no previous one and the buffer
is cleared. The same now goes for onexit. Also fixed some
potential uncaught exception situations in Pipe.